Only install oh-my-zsh for users that exist.
Not all items in the users data bag may actually be created as users. For example, recipe[users::sysadmins] only creates users in the sysadmins group, whereas other recipes may use other criteria to decide which users to create under which circumstances.

The recipes in this cookbook should therefore not assume that every item in the users data bag has been created as a user on the system. Doing so can cause these recipes to fail when they attempt to add files to the home directories of users that do not exist.